% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stored_data.R
\docType{data}
\name{data_tibble}
\alias{data_tibble}
\title{A Stored Tibble}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 12 columns.}
\usage{
data_tibble
}
\description{
The \code{mtcars} dataset with an additional list column (repeated) of \code{letters}.
}
\keyword{datasets}
