% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_locate.R
\name{dep_count_pkg}
\alias{dep_count_pkg}
\title{Count calls by package for a package}
\usage{
dep_count_pkg(pkg)
}
\arguments{
\item{pkg}{The package we should locale calls from}
}
\description{
Count calls by package for a package
}
