% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_locate.R
\name{dep_count_proj}
\alias{dep_count_proj}
\title{Count calls by package for a project}
\usage{
dep_count_proj(path = ".")
}
\arguments{
\item{path}{Project path to search in.}
}
\description{
Count calls by package for a project
}
