% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_locate.R
\name{dep_locate}
\alias{dep_locate}
\title{Locate calls to a particular dependency}
\usage{
dep_locate(pkg, path = ".")
}
\arguments{
\item{pkg}{The package we should locale calls from}

\item{path}{Project path to search in.}
}
\description{
Locate calls to a particular dependency
}
