% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dep_plot_time}
\alias{dep_plot_time}
\alias{dep_plot_size}
\alias{dep_plot_maintainer}
\title{Plot of dependency weights}
\usage{
dep_plot_time(pkg)

dep_plot_size(pkg)

dep_plot_maintainer(pkg)
}
\arguments{
\item{pkg}{package to plot}
}
\description{
`dep_plot_time()` plots the installation time to compile the source package.
`dep_plot_size()` plots the binary package size, `dep_plot_maintainer()`
plots counts of maintainers.
}
