% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_locate.R
\name{dep_usage_pkg}
\alias{dep_usage_pkg}
\title{Determine usage of depedencies for a package}
\usage{
dep_usage_pkg(pkg)
}
\arguments{
\item{pkg}{The package we should locale calls from}
}
\description{
Determine usage of depedencies for a package
}
