% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{dep_weight}
\alias{dep_weight}
\title{Calculate the package weight}
\usage{
dep_weight(packages, repos = c(CRAN = "https://cloud.r-project.org"),
  when = "last-week")
}
\arguments{
\item{packages}{Packages to weigh}

\item{repos}{The CRAN or CRAN-like repo to use}

\item{when}{\code{last_day}, \code{last_week} or \code{last_month}.
If this is given, then \code{from} and \code{to} are ignored.}
}
\value{
- funs - If you compare the # of exported funs vs the
  number of function you import / use in your package it gives you some idea
  of how much work it would take to remove the dependency. e.g. if you use 50
  functions from dplyr you would have to replicate 1/5 of more of the
  functionality from dplyr to remove the dependency, if you only use 1-5
  functions maybe it is more tractable

  But for packages like rcmdcheck or glue that only have a few functions you
  will have to replicate a larger proportion of the package even if you only
  import one function

  It also gives you some idea of the API size, which can give you an idea
  how likely it is to change in the future
}
\description{
Calculate the package weight
}
\examples{
\dontrun{
  dep_weight(c("dplyr", "data.table"))
}
}
