% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_locate.R
\name{dep_usage_proj}
\alias{dep_usage_proj}
\title{Determine usage of depedencies for a project}
\usage{
dep_usage_proj(path = ".")
}
\arguments{
\item{path}{Project path to search in.}
}
\description{
Determine usage of depedencies for a project
}
